requirements.txt
cachelib==0.13.0
cloudscraper==1.2.71
Flask==3.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
Werkzeug==3.1.3
